αλλαγες στο φυλλαδιο με τους καθετηρες συγχωνευση φυλλαδιων ξεχωριστα για οροφους και μεθ μικροδιορθωσεις

+++ app/src/main/java/micros_leader/george/nosileutikosfakelos/OROFOI/f_Isozigio_pros_apov/Isozigio_statheresDialog.java
@@ -153,7 +153,7 @@ public class Isozigio_statheresDialog extends DialogFragment {
 
             StringBuilder query = new StringBuilder() ;
 
-            int id = lista.get(i).getId();
+            long id = lista.get(i).getId();
             int itemID =  result.get(i).getItemID();
             String dateIN = result.get(i).getDatein();
